fix: article footer style
Update style to more clearly distinguish between article content and "metadata"

fix #1613
