+++ client/.storybook/preview.scss
@@ -26,7 +26,7 @@ body {
 
 /* Grids */
 .docs {
-  .reference-page {
+  .document-page {
     .page-header,
     .titlebar-container,
     .breadcrumb-container,
