+++ client/src/document/organisms/metadata/index.scss
@@ -1,0 +1,22 @@
+@import "~@mdn/minimalist/sass/vars/color-palette";
+@import "~@mdn/minimalist/sass/vars/layout";
+@import "~@mdn/minimalist/sass/vars/typography";
+
+.metadata {
+  background-color: $neutral-550;
+}
+
+.metadata-content-container {
+  margin: 0 auto;
+  max-width: $max-width-default;
+  padding: $base-spacing;
+
+  .on-github {
+    margin-bottom: $base-spacing;
+  }
+
+  .last-modified-date {
+    font-size: $tiny-text;
+    margin-bottom: 0;
+  }
+}
