+++ client/src/stories/grid/reference-page-grid.stories.mdx
@@ -5,7 +5,7 @@ import { Meta } from "@storybook/addon-docs/blocks";
 # Reference Page Grid
 
 <div class="docs">
-  <div class="page-wrapper reference-page">
+  <div class="page-wrapper document-page">
     <header class="page-header">
       <h1>Header</h1>
     </header>
